Decouple destination from interface
This PR adds the option to override/specify the DBUS destination of a
DBUS proxy for each created proxy object. This is necessary when the
same interface is used at multiple locations, where either to many
locations are possible to create individual object definitions for or
where the target is know at runtime only.

One such example is the evolution data server, which exposes the
datasource dynamically. A client contacts a fixed location (the
CalendarFactory) and opens the memo list there. At this moment EDS
creates a new dbus destination and returns the connection name and
objectpath for that destination. The client then binds that objectpath
and can issue methods of the org.gnome.evolution.dataserver.Calendar
interface on that.