Move EventloopGroup to connector level and properly close it
A trivial example like:

        try(DbusClient client = new DbusClient()) {
            client.connectSession();

            // here some interesting work could be done
        }

will block the JVM from exiting. The issue is, that once a connection is
established, a EventLoopGroup is created, that handles the communication
with netty. EventLoopGroups create Threads and these will block JVM
shutdown, unless the EventLoopGroups are shutdown.

While the DbusClient can be closed, this close is not passed on to the
underlying DbusConnector. This is fixed here and the asychronous
handling of network connections in netty is used and only a single
EventLoopGroup is created to handle all channels created.