Support Struct Binding and multiple returns
Members of structs in DBUS are addressed by their position and earch
member of a struct can have a different type. The approach taken here
is, that the structures positions are mapped to object properties. The
properties are expected to follow the java bean convention (i.e. there
has to be exactly one set<PropertyName> and one get<PropertyName> method
and the return type of the getter has to match the only parameter of the
setter).

This also allows to support multiple returns, as these are basicly
structs and thus can be trivially mapped.